Trying to get Confluence pickup mechanically-seeded DB connection to no avail.
If I start with confluence.cfg.xml taken from production, which has <setupStep>complete</setupStep>, Confluence fails to start citing it couldn't find the table named BANDANA.

If I set <setupStep> to some of the earlier phase values, such as setupdata-start, then it ignores properties that I seed in confluence.cfg.xml and demands the user to type it in.

On top of that, Confluence really appears unhappy unless the license is provided. The license is tied to server ID, which is generated fresh every time.

 So I'm giving up automating preseeding of confluence.cfg.xml. For now, let's have humans go through the setup wizard.