Disable LDAP cache.
One of the problems we found while fighting the spam is that even when
deleting an user from LDAP, the user in question can continue to spam
Wiki.

I think disabling LDAP cache in Confluence might solve this. At the same
time, this might have some unknown performance implication.

More information about this is in
https://confluence.atlassian.com/bamboo/configuring-the-caching-of-your-ldap-repository-289277216.html